Revert "feat[#7] 댓글 CRUD 구현"
