+++ src/main/java/com/lol/community/board/dto/request/BoardSearchRequest.java
@@ -24,10 +24,6 @@ public class BoardSearchRequest {
         return categories == 0;
     }
 
-    public boolean isEmptyGrade() {
-        return grades == 0;
-    }
-
     public boolean isExistGrade() {
         return grades != 0;
     }
