+++ src/main/java/com/lol/community/board/repository/BoardRepository.java
@@ -50,4 +50,5 @@ public interface BoardRepository extends JpaRepository<Board, Integer> {
             Pageable pageable
     );
 
+
 }
