--- src/main/java/com/lol/community/global/config/QuerydslConfig.java
@@ -1,18 +1,0 @@
-package com.lol.community.global.config;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.PersistenceContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class QuerydslConfig {
-  @PersistenceContext
-  private EntityManager entityManager;
-
-  @Bean
-  public JPAQueryFactory jpaQueryFactory() {
-    return new JPAQueryFactory(entityManager);
-  }
-}
