--- src/main/java/com/lol/community/reaction/domain/LikeReactionComment.java
@@ -1,45 +1,0 @@
-package com.lol.community.reaction.domain;
-
-import com.lol.community.comment.domain.Comment;
-import com.lol.community.global.BaseEntity;
-import com.lol.community.user.domain.User;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "comment_reaction")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class LikeReactionComment extends BaseEntity {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "id", updatable = false)
-  private Integer crId;
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "comment_id", nullable = false)
-  private Comment comment;
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "user_id", nullable = false)
-  private User user;
-
-  @Builder
-  public LikeReactionComment(Integer id, Comment comment, User user) {
-    this.crId = id;
-    this.comment = comment;
-    this.user = user;
-  }
-}
