+++ src/main/resources/application.yml
@@ -29,25 +29,3 @@ spring:
         show_sql: true
     database-platform: org.hibernate.dialect.MySQLDialect
     database: mysql
-
-#spring:
-#  application:
-#    name: lol-community
-#  datasource:
-#    driver-class-name: com.mysql.cj.jdbc.Driver
-#    url: jdbc:mysql://localhost:11802/lol
-#    username: root
-#    password: root
-#  jpa:
-#    hibernate:
-#      ddl-auto: update
-#    properties:
-#      hibernate:
-#        format_sql: true
-#        show_sql: true
-#  thymeleaf:
-#    cache: false
-
-#  spring.data.redis.host=127.0.0.1
-#  spring.data.redis.port=6379
-#  spring.data.redis.timeout=10
